Rename StatesRegistry to WillowRegistry
